+++ studio/src/app/modals/editor/app-publish/app-publish.tsx
@@ -1,7 +1,6 @@
 import {snapshotDeck, snapshotDoc} from '@deckdeckgo/sync';
 import {Component, Element, h, Listen, State} from '@stencil/core';
 import {AppIcon} from '../../../components/core/app-icon/app-icon';
-import {updatePublishedDeckOffline, updatePublishedDocOffline} from '../../../providers/publish/publish.provider';
 import editorStore from '../../../stores/editor.store';
 import i18n from '../../../stores/i18n.store';
 import {updateSlidesQRCode} from '../../../utils/editor/qrcode.utils';
@@ -18,12 +17,7 @@ export class AppPublish {
 
   private unsubscribeSnapshot: () => void | undefined;
 
-  private docListener;
-  private deckListener;
-
   async componentWillLoad() {
-    this.initOfflineUpdate();
-
     this.unsubscribeSnapshot = await this.initSnapshot();
   }
 
@@ -35,21 +29,8 @@ export class AppPublish {
     return snapshotDeck();
   }
 
-  /**
-   * We snapshot the doc/deck changes and have to replicate the value to the offline data to replicate the new "meta" information.
-   */
-  private initOfflineUpdate() {
-    this.docListener = editorStore.onChange('doc', updatePublishedDocOffline);
-    this.deckListener = editorStore.onChange('deck', updatePublishedDeckOffline);
-  }
-
   async componentDidLoad() {
     history.pushState({modal: true}, null);
-  }
-
-  disconnectedCallback() {
-    this.docListener?.();
-    this.deckListener?.();
   }
 
   @Listen('popstate', {target: 'window'})