+++ studio/src/app/providers/publish/publish.provider.ts
@@ -1,5 +1,4 @@
-import {Deck, Doc, Meta, throwError} from '@deckdeckgo/editor';
-import {updateOfflineDeck, updateOfflineDoc} from '@deckdeckgo/offline';
+import {Meta} from '@deckdeckgo/editor';
 import {publish as publishProvider, PublishInputs, publishUrl as publishUrlProvider} from '@deckdeckgo/sync';
 import {EnvironmentDeckDeckGoConfig} from '../../config/environment-config';
 import {EnvironmentConfigService} from '../../services/environment/environment-config.service';
@@ -18,28 +17,4 @@ export const publishUrl = async (meta: Meta | undefined): Promise<string> => {
   }
 
   return url;
-};
-
-export const updatePublishedDocOffline = async (doc: Doc | undefined) => {
-  if (!doc) {
-    return;
-  }
-
-  try {
-    await updateOfflineDoc(doc);
-  } catch (err) {
-    throwError(err);
-  }
-};
-
-export const updatePublishedDeckOffline = async (deck: Deck | undefined) => {
-  if (!deck) {
-    return;
-  }
-
-  try {
-    await updateOfflineDeck(deck);
-  } catch (err) {
-    throwError(err);
-  }
 };